requirements.txt
requests
beautifulsoup4
PyMuPDF
discord.py
Pillow